Revert "Revert "Issue#172""
