+++ akdb/src/opti/rel_eq_assoc.c
