+++ akdb/src/opti/rel_eq_projection_print_test.c
@@ -1,404 +1,7 @@
-/**
- @file rel_eq_projection.c Provides functions for for relational equivalences in projection
- */
-/*
- * This program is free software; you can redistribute it and/or modify
- * it under the terms of the GNU General Public License as published by
- * the Free Software Foundation; either version 2 of the License, or
- * (at your option) any later version.
- * 
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU Library General Public License for more details.
- * 
- * You should have received a copy of the GNU General Public License
- * along with this program; if not, write to the Free Software
- * Foundation, Inc., 51 Franklin Street, Fifth Floor Boston, MA 02110-1301,  USA
- */
-
-/**
- * @author Dino Laktašić.
- * ============> Optimization plan using Relational Algebra Equivalences <==============
- * Equivalence rule that apply on every equivalent expresion generated by Query optimizer
- * 
- * Rules to implement
- * Rule 1. projection comutes with selection that only uses attributes retained by the projection
- * 	p[L](s[L1](R)) = s[L1](p[L](R))
- * Rule 2. only the last in a sequence of projection operations is needed, the others can be omitted.
- *	p[L1](p[L2](...p[Ln](R)...)) = p[L1](R)
- * Rule 3a. distribution according to theta join, only if join includes attributes from L1 u L2
- *  p[L1 u L2](R1 t R2) = (p[L1](R1)) t (p[L2](R2))
- * Rule 3b. Let L1 u L2 be attributes from R1 and R2, respectively. Let L3 be attributes from R1, 
- * but are not in L1 u L2 and let L4 be attributes from R2, but are not in L1 u L2.
- *	p[L1 u L2](R1 t R2) = p[L1 u L2]((p[L1 u L3](R1)) t (p[L2 u L4](R2))) 
- * Rule 4. distribution according to union 
- *	p[L](R1 u R2) = (p[L](R1)) u (p[L](R2))
- * */
+//THIS FILE IS PART 2 OF REL_EQ_PROJECTION.C   ORIGINALLY IT WAS PART OF IT, BUT IN ORDER TO SPLIT THE CODE AND MAKE IT MORE READABLE IT MADE INTO 2 PARTS
 
 #include "rel_eq_projection.h"
 #include "../auxi/auxiliary.h"
-
-/**
- * @author Dino Laktašić.
- * @brief Function that checks if some set of attributes is subset of larger set, used in cascading of the projections
- * <ol>
- * <li>Tokenize set and subset of projection attributes and store each of them to it's own array</li>
- * <li>Check if the size of subset array is larger than the size of set array</li>
- * <li>if the subset array is larger return 0</li>
- * <li>else sort both arrays ascending</li>
- * <li>Compare the subset and set items at the same positions, starting from 0</li>
- * <li>if there is an item in the subset array that doesn't match attribute at the same position in the set array return 0</li>
- * <li>else continue comparing until final item in the subset array is ritched</li>
- * <li>on loop exit return EXIT_SUCCESS</li>
- * </ol>
- * @param list_elem_set first list element containing projection attributes 
- * @param list_elem_subset second list element containing projection attributes 
- * @return EXIT_SUCCESS if some set of attributes is subset of larger set, else returns EXIT_FAILURE
- */
-int AK_rel_eq_is_subset(struct list_node *list_elem_set, struct list_node *list_elem_subset) {
-    int len_set, len_subset, token_id = 0, set_id, subset_id;
-    char *temp_set, *temp_subset, *token_set, *token_subset, *save_token_set, *save_token_subset;
-    AK_PRO;
-    char *tokens_set[MAX_TOKENS] = {NULL};
-    char *tokens_subset[MAX_TOKENS] = {NULL};
-
-    len_set = len_subset = 0;
-
-    temp_set = (char *) AK_calloc(list_elem_set->size, sizeof (char));
-    temp_subset = (char *) AK_calloc(list_elem_subset->size, sizeof (char));
-
-    memcpy(temp_set, list_elem_set->data, list_elem_set->size);
-    memcpy(temp_subset, list_elem_subset->data, list_elem_subset->size);
-
-    AK_dbg_messg(HIGH, REL_EQ, "RULE - is (%s) subset of set (%s) in rel_eq_projection\n", list_elem_subset->data, list_elem_set->data);
-
-    for ((token_set = strtok_r(temp_set, ATTR_DELIMITER, &save_token_set)); token_set;
-            (token_set = strtok_r(NULL, ATTR_DELIMITER, &save_token_set)), token_id++) {
-        if (token_id < MAX_TOKENS - 1) {
-            tokens_set[token_id] = token_set;
-            len_set++;
-        }
-    }
-
-    token_id = 0;
-    for ((token_subset = strtok_r(temp_subset, ATTR_DELIMITER, &save_token_subset)); token_subset;
-            (token_subset = strtok_r(NULL, ATTR_DELIMITER, &save_token_subset)), token_id++) {
-        if (token_id < MAX_TOKENS - 1) {
-            tokens_subset[token_id] = token_subset;
-            len_subset++;
-        }
-    }
-
-    if (len_set < len_subset) {
-        AK_dbg_messg(HIGH, REL_EQ, "RULE - failed (%s) isn't subset of set (%s)!\n", list_elem_subset->data, list_elem_set->data);
-        AK_EPI;
-        return EXIT_FAILURE;
-    }
-
-    qsort(tokens_set, len_set, sizeof (char *), AK_strcmp);
-    qsort(tokens_subset, len_subset, sizeof (char *), AK_strcmp);
-
-    token_id = 0;
-
-    for (subset_id = 0; tokens_subset[subset_id] != NULL; subset_id++) {
-        for (set_id = 0; tokens_set[set_id] != NULL; set_id++) {
-            if (strcmp(tokens_set[set_id], tokens_subset[subset_id]) == 0) {
-                token_id++;
-            }
-        }
-    }
-
-    if (token_id != len_subset) {
-        AK_dbg_messg(HIGH, REL_EQ, "RULE - failed (%s) isn't subset of set (%s)!\n", list_elem_set->data, list_elem_subset->data);
-        AK_EPI;
-        return EXIT_FAILURE;
-    }
-
-    //May cause troubles
-    AK_free(temp_set);
-    AK_free(temp_subset);
-
-    AK_dbg_messg(HIGH, REL_EQ, "RULE - succeed (%s) is subset of set (%s).\n", list_elem_subset->data, list_elem_set->data);
-    AK_EPI;
-    return EXIT_SUCCESS;
-}
-
-/**
- * @author Dino Laktašić.
- * @brief Function that checks if selection uses only attributes retained by the projection before commuting
- * <ol>
- * <li>Tokenize set of projection attributes and store them to the array</li>
- * <li>For each attribute in selection condition check if exists in array of projection attributes</li>
- * <li>if exists increment match variable and break</li>
- * <li>else continue checking until the final attribute is checked</li>
- * <li>if match variable value equals 0 than return 0</li>
- * <li>else if match variable value greater than EXIT_SUCCESS, return EXIT_FAILURE</li>
- * </ol>
- * @param list_elem_attribs list element containing projection data 
- * @param list_elem_conds list element containing selection condition data
- * @return EXIT_SUCCESS if selection uses only attributes retained by projection, else returns EXIT_FAILURE
- */
-int AK_rel_eq_can_commute(struct list_node *list_elem_attribs, struct list_node *list_elem_conds) {
-    int next_chr, valid_cond_attribs, token_id = 0;
-    char *token_attr, *save_token_attr, *temp_attr, *temp_cond, *temp;
-    AK_PRO;
-    char *tokens[MAX_TOKENS] = {NULL};
-
-    AK_dbg_messg(HIGH, REL_EQ, "RULE - commute condition (%s) with projection (%s)\n", list_elem_conds->data, list_elem_attribs->data);
-
-    temp = (char *) AK_calloc(list_elem_conds->size, sizeof (char));
-    temp_cond = (char *) AK_calloc(list_elem_conds->size, sizeof (char));
-    temp_attr = (char *) AK_calloc(list_elem_attribs->size, sizeof (char));
-
-    memcpy(temp_attr, list_elem_attribs->data, list_elem_attribs->size);
-
-    for ((token_attr = strtok_r(temp_attr, ATTR_DELIMITER, &save_token_attr)); token_attr;
-            (token_attr = strtok_r(NULL, ATTR_DELIMITER, &save_token_attr)), token_id++) {
-        if (token_id < MAX_TOKENS - 1) {
-            tokens[token_id] = token_attr;
-        }
-    }
-
-    for (next_chr = 0; next_chr < list_elem_conds->size; next_chr++) {
-        if (list_elem_conds->data[next_chr] == ATTR_ESCAPE) {
-
-            memcpy(temp, list_elem_conds->data + next_chr + 1, list_elem_conds->size - next_chr);
-            memmove(temp_cond, temp, strcspn(temp, "`")); //ATTR_ESCAPE
-            next_chr += strlen(temp_cond) + 1;
-
-            valid_cond_attribs = 0;
-
-            for (token_id = 0; tokens[token_id] != NULL; token_id++) {
-
-                if (strcmp(temp_cond, tokens[token_id]) == 0) {
-                    valid_cond_attribs++;
-                    break;
-                }
-            }
-
-            if (valid_cond_attribs == 0) {
-                AK_dbg_messg(HIGH, REL_EQ, "RULE - commute condition with projection failed!\n");
-                AK_EPI;
-                return EXIT_FAILURE;
-            }
-        }
-    }
-
-    AK_free(temp);
-    AK_dbg_messg(HIGH, REL_EQ, "RULE - commute condition with projection succeed.\n");
-    AK_EPI;
-    return EXIT_SUCCESS;
-}
-
-/**
- * @author Dino Laktašić.
- * @brief Function that gets attributes for a given table and store them to the struct list_node 
- * <ol>
- * <li>Get the number of attributes in a given table</li>
- * <li>Get the table header for a given table</li>
- * <li>Initialize struct list_node</li>
- * <li>For each attribute in table header, insert attribute in struct list_node as new struct list_node element</li>
- * <li>return struct list_node</li>
- * </ol>
- * @param *tblName name of the table
- * @result struct list_node
- */
-struct list_node *AK_rel_eq_get_attributes(char *tblName) {
-    int next_attr;
-    int num_attr = AK_num_attr(tblName);
-    char *attr_name;
-    AK_PRO;
-
-    AK_header *table_header = (AK_header *) AK_get_header(tblName);
-
-    struct list_node *list_attr = (struct list_node *) AK_malloc(sizeof (struct list_node));
-    AK_Init_L3(&list_attr);
-
-    for (next_attr = 0; next_attr < num_attr; next_attr++) {
-        attr_name = (table_header + next_attr)->att_name;
-        AK_InsertAtEnd_L3(TYPE_ATTRIBS, attr_name, strlen(attr_name) + 1, list_attr);
-    }
-
-    attr_name = NULL;
-    AK_free(table_header);
-    AK_EPI;
-    return list_attr;
-}
-
-/* Another function to retreive and store table attributes, but in the array (may be useful somewhere later)
-const char *AK_rel_eq_get_attributes(char *tblName) {
-        int next_attr, num_attr = AK_num_attr(tblName);
-        char *attributes[MAX_ATTRIBUTES] = {NULL};
-        //char *attributes = (char *)AK_malloc(sizeof(MAX_ATTRIBUTES));
-	
-        //AK_header *table_header = (AK_header*)AK_calloc(num_attr, sizeof(AK_header));
-        AK_header *table_header = (AK_header *)AK_get_header(tblName);
-
-        for (next_attr = 0; next_attr < num_attr; next_attr++) {
-                attributes[next_attr] = (table_header + next_attr)->att_name;
-                //printf("attribute: %s\n", attributes[next_attr]);
-        }
-	
-        char *save_attributes = (char *)AK_malloc(strlen(*attributes) + 1);
-	
-        memcpy(save_attributes, attributes, sizeof(attributes));
-        return save_attributes;
-} */
-
-/**
- * @author Dino Laktašić.
- * @brief Function used for filtering and returning only those attributes from list of projection attributes that exist in the given table  
- * <ol>
- * <li>Get the attributes for a given table and store them to the AK_list</li>
- * <li>Tokenize set of projection attributes and store them to the array</li>
- * <li>For each attribute in the array check if exists in the previously created AK_list</li>
- * <li>if exists append attribute to the dynamic atributes char array</li>
- * <li>return pointer to char array with stored attribute/s</li>
- * </ol>
- * @param *attribs projection attributes delimited by ";" (ATTR_DELIMITER)
- * @param *tblName name of the table
- * @return filtered list of projection attributes as the AK_list
- */
-char *AK_rel_eq_projection_attributes(char *attribs, char *tblName) {
-    int len_tokens = 0;
-    int token_id = 0;
-    AK_PRO;
-    struct list_node *list_attr = AK_rel_eq_get_attributes(tblName);
-    struct list_node *list_el;
-
-    if (AK_Size_L2(list_attr) <= 0) {
-        printf("ERROR - table (%s) doesn't exists!\n", tblName);
-        AK_EPI;
-        return NULL;
-    }
-
-    char *token_attr, *save_token_attribs;
-    char *tokens_attribs[MAX_TOKENS] = {NULL};
-
-    char *ret_attributes = (char *) AK_calloc(MAX_VARCHAR_LENGTH, sizeof (char));
-    memset(ret_attributes, '\0', MAX_VARCHAR_LENGTH);
-
-    char *temp_attribs = (char *) AK_calloc(strlen(attribs), sizeof (char));
-
-    strcpy(temp_attribs, attribs);
-
-    AK_dbg_messg(HIGH, REL_EQ, "\nINPUT - attributes: (%s), tblName: (%s)\n", temp_attribs, tblName);
-
-    for ((token_attr = strtok_r(temp_attribs, ATTR_DELIMITER, &save_token_attribs)); token_attr;
-            (token_attr = strtok_r(NULL, ATTR_DELIMITER, &save_token_attribs)), token_id++) {
-        if (token_id < MAX_TOKENS - 1) {
-            tokens_attribs[len_tokens] = token_attr;
-            len_tokens++;
-        }
-    }
-
-    for (token_id = 0; token_id < len_tokens; token_id++) {
-        list_el = (struct list_node *) AK_First_L2(list_attr);
-
-        while (list_el) {
-            if (strcmp(list_el->data, tokens_attribs[token_id]) == 0) {
-                if (strlen(ret_attributes) > 0) {
-                    strcat(ret_attributes, ATTR_DELIMITER);
-                }
-                strcat(ret_attributes, list_el->data);
-            }
-            list_el = list_el->next;
-        }
-    }
-
-    AK_DeleteAll_L3(&list_attr);
-    AK_free(temp_attribs);
-
-    AK_dbg_messg(HIGH, REL_EQ, "RETURN - attributes for new projection (%s)\n", ret_attributes);
-    AK_EPI;
-    return ret_attributes;
-}
-
-/**
- * @author Dino Laktašić.
- * @brief Function used for filtering and returning only attributes from selection or theta_join condition 
- * @param list_elem list element that contains selection or theta_join condition data
- * @return only attributes from selection or theta_join condition as the AK_list
- */
-char *AK_rel_eq_collect_cond_attributes(struct list_node *list_elem) {
-    int next_chr = 0;
-    int next_address = 0;
-    int attr_end = -1;
-    AK_PRO;
-    char *temp_cond = (char *) AK_malloc(list_elem->size);
-    strcpy(temp_cond, list_elem->data);
-
-    char *attr = (char *) AK_calloc(MAX_VARCHAR_LENGTH, sizeof (char));
-
-    while (next_chr < list_elem->size) {
-        if (temp_cond[next_chr] == ATTR_ESCAPE) { //'`'
-            next_chr++;
-            if (++attr_end) {
-                attr_end = -1;
-            } else {
-                if (next_address > 0) {
-                    strcpy(attr + next_address++, ATTR_DELIMITER);
-                }
-            }
-        }
-
-        if (!attr_end) {
-            strcpy(attr + next_address++, &temp_cond[next_chr]);
-        }
-        next_chr++;
-    }
-
-    AK_free(temp_cond);
-    strcpy(attr + next_address, "\0");
-    AK_EPI;
-    return attr;
-}
-
-/**
- * @author Dino Laktašić.
- * @brief Function which removes duplicate attributes from attributes expresion 
- * @param *attribs attributes from which to remove duplicates
- * @return pointer to char array without duplicate attributes
- */
-char *AK_rel_eq_remove_duplicates(char *attribs) {
-    int next_attr = 0;
-    int token_id = 0;
-    int exist_attr;
-    AK_PRO;
-    char *temp_attribs, *token_attr, *save_token_attribs;
-    char *tokens_attribs[MAX_TOKENS] = {NULL};
-    char *attr = (char *) AK_calloc(MAX_VARCHAR_LENGTH, sizeof (char));
-    memset(attr, '\0', MAX_VARCHAR_LENGTH);
-
-    temp_attribs = (char *) AK_calloc(strlen(attribs) + 1, sizeof (char));
-    memcpy(temp_attribs, attribs, strlen(attribs));
-    memcpy(temp_attribs + strlen(attribs) + 1, "\0", 1);
-
-    for ((token_attr = strtok_r(temp_attribs, ATTR_DELIMITER, &save_token_attribs)); token_attr;
-            (token_attr = strtok_r(NULL, ATTR_DELIMITER, &save_token_attribs)), token_id++) {
-        if (token_id < MAX_TOKENS - 1) {
-            tokens_attribs[token_id] = token_attr;
-            exist_attr = 0;
-
-            for (next_attr = 0; next_attr < token_id; next_attr++) {
-                if (strcmp(tokens_attribs[next_attr], token_attr) == 0) {
-                    exist_attr = 1;
-                }
-            }
-
-            if (!exist_attr) {
-                if (token_id > 0) {
-                    strcat(attr, ATTR_DELIMITER);
-                }
-                strcat(attr, token_attr);
-            }
-        }
-    }
-
-    AK_EPI;
-    return attr;
-}
 
 /**
  * @author Dino Laktašić.
